+++ lib/shoehorn/overlay.ex
@@ -1,15 +1,15 @@
-defmodule Bootloader.Overlay do
-  alias Bootloader.Utils
+defmodule Shoehorn.Overlay do
+  alias Shoehorn.Utils
 
   defstruct [hash: nil, delta: []]
 
   @type t :: %__MODULE__{
     hash: String.t,
-    delta: [Bootloader.Application.t]
+    delta: [Shoehorn.Application.t]
   }
 
   def load(sources, targets) when is_list(sources) and is_list(targets) do
-    delta = Bootloader.Application.compare(sources, targets)
+    delta = Shoehorn.Application.compare(sources, targets)
     %__MODULE__{
       delta: delta
     }
@@ -18,7 +18,7 @@ defmodule Bootloader.Overlay do
 
   def apply(%__MODULE__{} = overlay, overlay_dir) do
     overlay_path = Path.join(overlay_dir, hash(overlay.delta))
-    Enum.each(overlay.delta, &Bootloader.Application.apply(&1, overlay_path))
+    Enum.each(overlay.delta, &Shoehorn.Application.apply(&1, overlay_path))
   end
 
   def hash(applications) do