+++ test/fixtures/fail_init/config/config.exs
@@ -1,3 +1,3 @@
 use Mix.Config
 
-config :bootloader, app: :fail_init
+config :shoehorn, app: :fail_init